requirements.dev.txt
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.4
fire==0.6.0
fonttools==4.51.0
fsspec==2024.3.1
imageio==2.34.0
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.8.4
medmnist==3.0.1
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-image==0.23.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sympy==1.12
termcolor==2.4.0
threadpoolctl==3.4.0
tifffile==2024.2.12
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
